+++ types/milesight/EM300-TH/uplink.spec.js
@@ -7,6 +7,7 @@ const { assert } = chai;
 
 describe("AM104 Uplink", () => {
   let defaultSchema = null;
+  let lifecycleSchema = null;
   let consume = null;
   before((done) => {
     const script = rewire("./uplink.js");
@@ -15,6 +16,17 @@ describe("AM104 Uplink", () => {
       .loadSchema(`${__dirname}/default.schema.json`)
       .then((parsedSchema) => {
         defaultSchema = parsedSchema;
+        done();
+      });
+  });
+
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/lifecycle.schema.json`)
+      .then((parsedSchema) => {
+        lifecycleSchema = parsedSchema;
         done();
       });
   });
@@ -33,8 +45,18 @@ describe("AM104 Uplink", () => {
         assert.isNotNull(value);
         assert.typeOf(value.data, "object");
 
-        if (value.topic === "default") {
+        if (value.topic === "lifecycle") {
           assert.equal(value.data.batteryLevel, 100);
+          validate(value.data, lifecycleSchema, { throwError: true });
+        }
+      });
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        if (value.topic === "default") {
           assert.equal(value.data.temperature, 27.2);
           assert.equal(value.data.humidity, 56.5);